+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Update to Grafana 9.3.6 (#147)
 - Update CI and Release workflows (#147)
 - Update README to include JavaScript tutorial (#148)
+- Add custom styles (#149)
 
 ## 2.2.0 (2023-01-09)